+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - `Innmind\Server\Status\Exception\EmptyPathNotAllowed`
 - `Innmind\Server\Status\Exception\BytesCannotBeNegative`
 - `Innmind\Server\Status\Exception\EmptyCommandNotAllowed`
+- `Innmind\Server\Status\Exception\OutOfBoundsPercentage`
 
 ## 4.1.1 - 2024-09-30